misc(lib/error): move IpmiError into its own module
